🦈🏠❗ Well I did things - Hopefully nothing breaks ❗🏠🦈

+++ src/os/mac/create_symbolic_links.sh
@@ -1,59 +1,47 @@
 #!/usr/bin/env bash
 
-cd "$(dirname "${BASH_SOURCE[0]}")" &&
-  . "../utils.sh"
-
+cd "$(dirname "${BASH_SOURCE[0]}")" && . "../utils.sh"
 srcdir="$(cd .. && pwd)"
-
-backups="${backupsdir:-$HOME/.local/backups/dotfiles/desktop}"
+backups="${backupsdir:-$HOME/.local/backups/os}"
 mkdir -p "$backups"/{configs,home}
 
-if [ ! -f "$HOME/.ncmpcpp/config" ]; then
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+if [ ! -f "$HOME/.ncmpcpp/config" ] && [ -f "$srcdir/config/mpd/ncmpcpp.conf" ]; then
   mkdir "$HOME/.ncmpcpp"
   ln -sf "$srcdir/config/mpd/ncmpcpp.conf" "$HOME/.ncmpcpp/config"
 fi
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 declare -a FILES_TO_SYMLINK=(
-  \
-  "config/bash/bash_logout"
-  "config/bash/bash_profile"
-  "config/bash/bashrc"
-  "shell/curlrc"
-  "shell/dircolors"
-  "shell/inputrc"
-  "shell/profile"
-  "shell/screenrc"
-  "shell/wgetrc"
-
+  # "config/bash/bash_logout"
+  # "config/bash/bash_profile"
+  # "config/bash/bashrc"
+  # "shell/curlrc"
+  # "shell/dircolors"
+  # "shell/inputrc"
+  # "shell/profile"
+  # "shell/screenrc"
+  # "shell/wgetrc"
+  "shell/face"
 )
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 declare -a CONFFOLDERS_TO_SYMLINK=(
-  \
-  "bash"
-  "dircolors"
-  "filezilla"
-  "fish"
-  "fontconfig"
-  "git"
-  "mpd"
-  "nano"
-  "neofetch"
-  "neovim"
-  "tmux"
-  "vim"
-  "youtube-dl"
-  "zsh"
-
+  # "bash"
+  # "dircolors"
+  # "filezilla"
+  # "fish"
+  # "fontconfig"
+  # "git"
+  # "mpd"
+  # "nano"
+  # "neofetch"
+  # "neovim"
+  # "tmux"
+  # "vim"
+  # "youtube-dl"
+  # "zsh"
 )
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 declare -a LIBRARYFILES_TO_SYMLINK=(
-  \
   "Developer/Xcode/UserData/FontAndColorThemes/Dracula.xccolortheme"
   "Preferences/com.apple.systempreferences.plist"
   "Preferences/com.googlecode.iterm2.plist"
@@ -61,24 +49,18 @@ declare -a LIBRARYFILES_TO_SYMLINK=(
   "Preferences/com.apple.Terminal.plist"
   "Preferences/com.apple.finder.plist"
   "Preferences/org.m0k.transmission.plist"
-
 )
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 backup_symlinks() {
   local i=""
   local sourceFile=""
   local targetFile=""
   local skipQuestions=true
-
   skip_questions "$@" && skipQuestions=true
-
   for i in "${FILES_TO_SYMLINK[@]}"; do
     sourceFile="$srcdir/$i"
     targetFile="$HOME/.$(printf "%s" "$i" | sed "s/.*\/\(.*\)/\1/g")"
     nameFile="$(printf "%s" "$i" | sed "s/.*\/\(.*\)/\1/g")"
-
     if [ -f "$targetFile" ] && [ ! -L "$targetFile" ] && [ -e "$sourceFile" ]; then
       execute \
         "mv -f $targetFile $backups/home/$nameFile" \
@@ -86,22 +68,17 @@ backup_symlinks() {
     fi
   done
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 backup_confsymlinks() {
   local i=""
   local sourceFile=""
   local targetFile=""
   local skipQuestions=true
-
   skip_questions "$@" && skipQuestions=true
-
   for i in "${CONFFOLDERS_TO_SYMLINK[@]}"; do
     sourceFile="$srcdir/config/$i"
     targetFile="$HOME/.config/$i"
     nameFile="$(printf "%s" "$i" | sed "s/.*\/\(.*\)/\1/g")"
-
     if [ -f "$targetFile" ] || [ -d "$targetFile" ] && [ ! -L "$targetFile" ] && [ ! -f "$srcdir/config/$i/install.sh" ] && [ -e "$sourceFile" ]; then
       execute \
         "mv -f $targetFile $backups/home/config/$nameFile" \
@@ -109,48 +86,36 @@ backup_confsymlinks() {
     fi
   done
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 backup_librarysymlinks() {
   local i=""
   local sourceFile=""
   local targetFile=""
   local skipQuestions=true
-
-  skip_questions "$@" &&
-    skipQuestions=true
-
+  skip_questions "$@" && skipQuestions=true
   for i in "${LIBRARYFILES_TO_SYMLINK[@]}"; do
     sourceFile="$srcdir/Library⁩/$i"
     targetFile="$HOME/Library⁩/$i"
     nameFile="$(printf "%s" "$i" | sed "s/.*\/\(.*\)/\1/g")"
-
     if [ -f "$targetFile" ] || [ -d "$targetFile" ] && [ ! -L "$targetFile" ]; then
       execute \
         "rsync -aq $targetFile/. $backups/home/Library⁩/$nameFile >/dev/null 2>&1 || true" \
         "Backing up $targetFile → $backups/home/Library⁩/$nameFile"
     fi
-
   done
-
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 create_symlinks() {
   setup() {
     unlink "$targetFile" 2>/dev/null || rm -Rf "$targetFile" 2>/dev/null
     ln -sf "$sourceFile" "$targetFile"
   }
-
   print_in_purple "\n   • Create file symlinks\n"
   local i=""
   local sourceFile=""
   local targetFile=""
   local skipQuestions=true
   skip_questions "$@" && skipQuestions=true
-
   for i in "${FILES_TO_SYMLINK[@]}"; do
     sourceFile="$srcdir/$i"
     targetFile="$HOME/.$(printf "%s" "$i" | sed "s/.*\/\(.*\)/\1/g")"
@@ -161,22 +126,18 @@ create_symlinks() {
     fi
   done
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 create_confsymlinks() {
   setup() {
     unlink "$targetFile" 2>/dev/null || rm -Rf "$targetFile" 2>/dev/null
     ln -sf "$sourceFile" "$targetFile"
   }
-
   print_in_purple "\n   • Create config symlinks\n"
   local i=""
   local sourceFile=""
   local targetFile=""
   local skipQuestions=true
   skip_questions "$@" && skipQuestions=true
-
   for i in "${CONFFOLDERS_TO_SYMLINK[@]}"; do
     sourceFile="$srcdir/config/$i"
     targetFile="$HOME/.config/$i"
@@ -187,9 +148,7 @@ create_confsymlinks() {
     fi
   done
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 create_librarysymlinks() {
   if [[ "$OSTYPE" =~ ^darwin ]]; then
     print_in_purple "\n   • Create library symlinks\n"
@@ -197,16 +156,12 @@ create_librarysymlinks() {
     local sourceFile=""
     local targetFile=""
     local skipQuestions=true
-
-    skip_questions "$@" &&
-      skipQuestions=true
-
+    skip_questions "$@" skipQuestions=true
     for i in "${LIBRARYFILES_TO_SYMLINK[@]}"; do
       sourceFile="$srcdir/Library/$i"
       targetFile="$HOME/Library/$i"
       mkdir -p ${targetFile%/*}
       rm -Rf "$targetFile"
-
       if [ ! -e "$targetFile" ]; then
         execute \
           "ln -sf $sourceFile $targetFile" \
@@ -219,10 +174,8 @@ create_librarysymlinks() {
     done
   fi
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 main() {
-
   backup_symlinks "$@"
   backup_confsymlinks "$@"
   backup_librarysymlinks "$@"
@@ -230,6 +183,7 @@ main() {
   create_confsymlinks "$@"
   create_librarysymlinks "$@"
 }
-
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 main "$@"
-unset main
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+# end
