+++ src/os/mac/install_tmux.sh
@@ -1,36 +1,31 @@
 #!/usr/bin/env bash
 
-cd "$(dirname "${BASH_SOURCE[0]}")" &&
-  . "../utils.sh"
-
+cd "$(dirname "${BASH_SOURCE[0]}")" && . "../utils.sh"
 srcdir="$(cd .. && pwd)"
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 setup_tmux() {
+  __cmd_exists dfmgr && dfmgr install tmux && return
+  [ -d "$srcdir/config/tmux" ] || return
   if [ -f "$srcdir/config/tmux/install.sh" ]; then
     execute "bash -c $srcdir/config/tmux/install.sh" "Installing tmux: $srcdir/config/tmux/install.sh"
-
   elif [ -d "$srcdir/config/tmux" ]; then
     if [ -L ~/.tmux ]; then unlink ~/.tmux 2>/dev/null; fi
     if [ -d ~/.tmux ]; then rm -Rf -f ~/.tmux 2>/dev/null; fi
-
     execute \
       "ln -sf ~/.local/dotfiles/src/config/tmux ~/.tmux" \
       "$srcdir/config/tmux → ~/.tmux"
-
     execute \
       "ln -sf $srcdir/config/tmux/tmux.conf ~/.tmux.conf" \
       "$srcdir/config/tmux/tmux.conf → ~/.tmux.conf"
   else
     exit
   fi
-
 }
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 setup_tmuxplugins() {
+  [ -d "$srcdir/config/tmux" ] || return
   if [ -d "$HOME/.local/share/tmux/tpm/tpm" ] && [ ! -f "$srcdir/config/tmux/install.sh" ]; then
     execute \
       "git -C $HOME/.local/share/tmux/tpm/tpm pull -q" \
@@ -50,16 +45,12 @@ setup_tmuxplugins() {
       "Installing tmux plugins → $HOME/.local/share/tmux/tpm"
   fi
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 main() {
-
   setup_tmux
   setup_tmuxplugins
-
 }
-
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 main
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+# end
